:recycle: Refactor: expect/actual 코드 core/di -> core/platform으로 이동
