+++ composeApp/src/androidMain/kotlin/com/dh/ondot/core/platform/KakaoSignIn.kt
@@ -1,4 +1,4 @@
-package com.dh.ondot.core.di
+package com.dh.ondot.core.platform
 
 import android.content.Context
 import android.util.Log
