+++ composeApp/src/commonMain/kotlin/com/dh/ondot/core/platform/AppModule.kt
@@ -1,13 +1,13 @@
-package com.dh.ondot.core.di
+package com.dh.ondot.core.platform
 
+import com.dh.ondot.core.network.NetworkClient
+import com.dh.ondot.core.network.TokenProvider
 import com.dh.ondot.data.repository.MemberRepositoryImpl
 import com.dh.ondot.data.repository.PlaceRepositoryImpl
 import com.dh.ondot.data.repository.ScheduleRepositoryImpl
 import com.dh.ondot.domain.repository.MemberRepository
 import com.dh.ondot.domain.repository.PlaceRepository
 import com.dh.ondot.domain.repository.ScheduleRepository
-import com.dh.ondot.core.network.NetworkClient
-import com.dh.ondot.core.network.TokenProvider
 import org.koin.core.module.Module
 import org.koin.dsl.module
