+++ composeApp/src/iosMain/kotlin/com/dh/ondot/core/platform/AppleSignIn.kt
@@ -1,4 +1,4 @@
-package com.dh.ondot.core.di
+package com.dh.ondot.core.platform
 
 import kotlinx.cinterop.BetaInteropApi
 import platform.AuthenticationServices.ASAuthorization
